Move SyslogRecordFactory to its own file
